move out of common some no available function in hw for libskycoin_assert ref #108
